+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Possible log types:
 
 - [added] Added the `HumitidySensor` and `PowerConsumptionSensor` sensor types.
 - [added] Added the `timezone` key under `location` section.
+- [added] Added the `links` section.
 
 ### v0.8.1 (2021-07-06)